Implement isaacParsonsQuestion
