+++ src/components/semantic/styles/semantic.module.css
@@ -93,12 +93,12 @@
 }
 
 .boxUp {
-    right: 66px;
+    right: 56px;
     top: -9px;
 }
 
 .boxDown {
-    right: 31px;
+    right: 26px;
     top: -9px;
 }
